Move getting-started to a server page w Client component
